refactor: extract plan_execute into its own subpackage
Move executor, planner, runner, and models (Plan, PlanStep, StepResult)
into src/agent/plan_execute/. AgentRunner ABC and OrchestratorResult
remain in src/agent/ as the shared orchestration contract. Update all
imports and patch targets in tests accordingly.
