+++ src/agent/__init__.py
@@ -1,7 +1,9 @@
 """MCP plan-execute orchestration package."""
 
-from .runner import AgentRunner, PlanExecuteRunner
-from .models import OrchestratorResult, Plan, PlanStep, StepResult
+from .runner import AgentRunner
+from .plan_execute.runner import PlanExecuteRunner
+from .models import OrchestratorResult
+from .plan_execute.models import Plan, PlanStep, StepResult
 
 __all__ = [
     "AgentRunner",
