+++ src/agent/runner.py
@@ -1,41 +1,13 @@
-"""Entry-point runner for the plan-execute workflow using MCP servers.
-
-Replaces AgentHive's combination of PlanningWorkflow + SequentialWorkflow with
-an MCP-native implementation:
-
-  AgentHive                       plan_execute
-  ────────────────────────────    ─────────────────────────────
-  PlanningWorkflow.generate_steps → Planner.generate_plan
-  SequentialWorkflow.run          → Executor.execute_plan
-  ReactAgent.execute_task         → _list_tools + _call_tool (MCP stdio)
-"""
+"""Abstract base class for all agent runners."""
 
 from __future__ import annotations
 
-import logging
 from abc import ABC, abstractmethod
 from pathlib import Path
 
 from llm import LLMBackend
 
-_log = logging.getLogger(__name__)
-
-from .executor import Executor
 from .models import OrchestratorResult
-from .planner import Planner
-
-_SUMMARIZE_PROMPT = """\
-You are summarizing the results of a multi-step task execution for an \
-industrial asset operations system.
-
-Original question: {question}
-
-Step-by-step execution results:
-{results}
-
-Provide a concise, direct answer to the original question based on the results
-above. Do not repeat the individual steps — just give the final answer.
-"""
 
 
 class AgentRunner(ABC):
@@ -57,79 +29,3 @@ class AgentRunner(ABC):
     @abstractmethod
     async def run(self, question: str) -> OrchestratorResult:
         """Run the agent on *question* and return a structured result."""
-
-
-class PlanExecuteRunner(AgentRunner):
-    """Entry-point for plan-and-execute workflows using MCP servers as tool providers.
-
-    Usage::
-
-        from plan_execute import PlanExecuteRunner
-        from llm import LiteLLMBackend
-
-        runner = PlanExecuteRunner(llm=LiteLLMBackend("watsonx/meta-llama/llama-3-3-70b-instruct"))
-        result = await runner.run("What are the assets at site MAIN?")
-        print(result.answer)
-
-    Args:
-        llm: LLM backend used for planning, tool selection, and summarisation.
-        server_paths: Override MCP server specs.  Keys must match the server
-                      names the planner will assign steps to.  Values are
-                      either a uv entry-point name (str) or a Path to a
-                      script file.  Defaults to all five registered servers.
-    """
-
-    def __init__(
-        self,
-        llm: LLMBackend,
-        server_paths: dict[str, Path | str] | None = None,
-    ) -> None:
-        super().__init__(llm, server_paths)
-        self._planner = Planner(llm)
-        self._executor = Executor(llm, server_paths)
-
-    async def run(self, question: str) -> OrchestratorResult:
-        """Run the full plan-execute loop for a question.
-
-        Steps:
-          1. Discover available servers from registered MCP servers.
-          2. Use the LLM to decompose the question into an execution plan.
-          3. Execute each plan step by routing tool calls to MCP servers.
-          4. Summarise the step results into a final answer.
-
-        Args:
-            question: The user question to answer.
-
-        Returns:
-            OrchestratorResult with the final answer, the generated plan, and
-            the per-step execution history.
-        """
-        # 1. Discover
-        _log.info("Discovering server capabilities...")
-        server_descriptions = await self._executor.get_server_descriptions()
-
-        # 2. Plan
-        _log.info("Planning...")
-        plan = self._planner.generate_plan(question, server_descriptions)
-        _log.info("Plan has %d step(s).", len(plan.steps))
-
-        # 3. Execute
-        history = await self._executor.execute_plan(plan, question)
-
-        # 4. Summarise
-        _log.info("Summarising...")
-        results_text = "\n\n".join(
-            f"Step {r.step_number} — {r.task} (server: {r.server}):\n"
-            + (r.response if r.success else f"ERROR: {r.error}")
-            for r in history
-        )
-        answer = self._llm.generate(
-            _SUMMARIZE_PROMPT.format(question=question, results=results_text)
-        )
-
-        return OrchestratorResult(
-            question=question,
-            answer=answer,
-            plan=plan,
-            history=history,
-        )
