+++ src/agent/tests/test_models.py
@@ -1,6 +1,6 @@
 """Tests for Plan and StepResult data models."""
 
-from agent.models import Plan, PlanStep, StepResult
+from agent.plan_execute.models import Plan, PlanStep, StepResult
 
 
 def _step(n: int, deps: list[int] | None = None) -> PlanStep:
